Factor out job creation success test
